// src/api/types/operations-aor-details-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Piattaforma Notifiche: API del BFF dei servizi RADD
 * API utilizzate dal Backend for Frontend dei servizi RADD. 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { OperationAorDetailResponse } from './operation-aor-detail-response';
// May contain unused imports in some cases
// @ts-ignore
import { OperationResponseStatus } from './operation-response-status';

/**
 * Id pratiche recuperate
 * @export
 * @interface OperationsAorDetailsResponse
 */
export interface OperationsAorDetailsResponse {
    /**
     * 
     * @type {Array<OperationAorDetailResponse>}
     * @memberof OperationsAorDetailsResponse
     */
    'elements'?: Array<OperationAorDetailResponse>;
    /**
     * 
     * @type {boolean}
     * @memberof OperationsAorDetailsResponse
     */
    'result'?: boolean;
    /**
     * 
     * @type {OperationResponseStatus}
     * @memberof OperationsAorDetailsResponse
     */
    'status'?: OperationResponseStatus;
}